requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
playwright>=1.40.0
scikit-learn>=1.4.0
rake-nltk>=1.0.6
nltk>=3.8.1
pandas>=2.1.0
